chore: Improve health check result accuracy and contracts (#1334)
* chore: Improve health check result accuracy and contracts

- Health checks now return Unhealthy with clear messages when core operations fail (e.g., missing resources, failed connections), instead of generic or Degraded results.
- Standardized use of (isTimelyResponse, result) tuple for timeout and success/failure handling.
- Changed CouchDbOptions and MongoDbOptions CommandAsync to return Task<bool> for consistent success indication.
- PubSubOptions now requires ProjectName; health check validates and uses it.
- LiteDBOptions.Timeout is now int for consistency.
- Updated tests to match new result semantics and improved mocking.
- Refactored code for clarity, error handling, and style.
- Public API updated for new CommandAsync signatures.

* fix: Updated Public API Definition

* fix: Nullable ProjectName

* chore(pulsar): Updated HealthCheck to validate unhealthy states

* fix(bigquery): Updated GCP.BigQuery

* fix(bigtable): Modernized BigTable

+++ src/NetEvolve.HealthChecks.LiteDB/LiteDBOptions.cs
@@ -27,5 +27,5 @@ public record LiteDBOptions
     /// <value>
     /// The timeout in milliseconds. Default value is 100 milliseconds.
     /// </value>
-    public double Timeout { get; set; } = 100;
+    public int Timeout { get; set; } = 100;
 }
